+++ src/test/resources/features/middlemanUser.feature
@@ -26,10 +26,10 @@ Feature: Login
       |name    |email                |phone        |address     |
       |toko ernawt12|ernatoko1@gmail.com|081234566661|jalan cinta|
 
+
   @user
   Scenario Outline: Update user without properti name
     Given user click button profile
-#    And user can see pop up menu
     And user choose button my profile
     Then user will be redirect to page my profile
     And user click button edit
@@ -37,19 +37,81 @@ Feature: Login
     And user input name "<name>"
     And user input email1 "<email>"
     And user input phone number1 "<phone>"
-#    And user input password1 "<password>"
+    And user input address1 "<address>"
+    And user click edit button
+    Then user see required alert on username field
+    Examples:
+      |name   |email              |phone       |address   |
+      |       |ernatoko1@gmail.com|081234566661|jalan cinta|
+
+  @user
+  Scenario Outline: Update user with already registered email
+    Given user click button profile
+    And user choose button my profile
+    Then user will be redirect to page my profile
+    And user click button edit
+    And user can see pop up edit profile
+    And user input name "<name>"
+    And user input email1 "<email>"
+    And user input phone number1 "<phone>"
+    And user input address1 "<address>"
+    And user click edit button
+    Then user see required alert on username field
+    Examples:
+      |name       |email                  |phone      |address     |
+      |toko coba2 |tokocobalagi1@gmail.com|12345773334|jalan cinta|
+
+
+  @user
+  Scenario Outline: Update user with email not valid
+    Given user click button profile
+    And user choose button my profile
+    Then user will be redirect to page my profile
+    And user click button edit
+    And user can see pop up edit profile
+    And user input name "<name>"
+    And user input email1 "<email>"
+    And user input phone number1 "<phone>"
+    And user input address1 "<address>"
+    And user click edit button
+    Then user see required alert on username field
+    Examples:
+      |name    |email            |phone      |address     |
+      |toko erna123|tokoernagmail|12345444444|jalan cinta|
+
+  @user
+  Scenario Outline: Update user with phone number not valid
+    Given user click button profile
+    And user choose button my profile
+    Then user will be redirect to page my profile
+    And user click button edit
+    And user can see pop up edit profile
+    And user input name "<name>"
+    And user input email1 "<email>"
+    And user input phone number1 "<phone>"
     And user input address1 "<address>"
     And user click edit button
     Then user see required alert on username field
     Examples:
       |name    |email                |phone        |address     |
-      |       |ernatoko1@gmail.com|081234566661|jalan cinta|
+      |toko erna123|ernatoko1@gmail.com|1234|jalan cinta|
 
-
-
-
-
-
+  @user
+  Scenario Outline: Update user with empty data on json file
+    Given user click button profile
+    And user choose button my profile
+    Then user will be redirect to page my profile
+    And user click button edit
+    And user can see pop up edit profile
+    And user input name "<name>"
+    And user input email1 "<email>"
+    And user input phone number1 "<phone>"
+    And user input address1 "<address>"
+    And user click edit button
+    Then user see required alert on username field
+    Examples:
+      |name    |email                |phone        |address     |
+      |        |                      |              |           |
